requirements.txt
rich==13.7.0
python-dotenv==1.0.0
yara-python==4.3.1
requests==2.31.0
